fix: #137 refactor search to use query string instead of path variable (no more weired jumping and routing); fix collapse to stay open until closed or clicked outside;
