+++ frontend/src/assets/styles/styles.scss
@@ -335,3 +335,7 @@ main {
 		margin-right: 0;
 	}
 }
+
+.cursor-pointer {
+	cursor: pointer;
+}
