--- frontend/src/pages/search/_tags/index.vue
@@ -1,5 +1,0 @@
-<script>
-import Search from '~/pages/search/_tags/_term';
-
-export default Search;
-</script>
